+++ src/flowcean/core/environment/__init__.py
@@ -5,7 +5,7 @@ __all__ = [
     "NotLoadedError",
     "JoinedEnvironment",
     "OfflineEnvironment",
-    "StackEnvironment",
+    "ChainEnvironment",
     "StreamingOfflineData",
     "TransformedEnvironment",
     "TransformedEnvironment",
@@ -13,9 +13,9 @@ __all__ = [
 
 from .active import ActiveEnvironment
 from .base import Environment, NotLoadedError
+from .chain import ChainEnvironment
 from .incremental import IncrementalEnvironment
 from .joined import JoinedEnvironment
 from .offline import OfflineEnvironment
-from .stack import StackEnvironment
 from .streaming import StreamingOfflineData
 from .transformed import TransformedEnvironment
